google_calendar_mcp/requirements.txt
mcp
httpx
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
